feat: move datawave conditions to out of the spark3
