+++ spark3/ethereum/condition.py
@@ -1,12 +1,9 @@
 from typing import Callable, TypeVar, Generic, Dict, Optional
 
-from eth_utils import encode_hex, event_abi_to_log_topic
-from eth_utils import function_abi_to_4byte_selector
 from pyspark.sql import DataFrame
 from pyspark.sql.functions import col
 
 from spark3.exceptions import ColumnNotFoundInDataFrame
-from spark3.utils import hash_unsafe_bytes
 
 T = TypeVar('T')
 
@@ -23,14 +20,6 @@ class Condition(Generic[T]):
             raise ColumnNotFoundInDataFrame(f'{self.key_alias}({self.key_type})', df)
 
         return df.filter(col(self.key_alias) == self.transform(raw))
-
-
-def return_self_condition(alias: str, _type: str) -> Condition[str]:
-    return Condition(key_alias=alias, key_type=_type, transform=lambda x: x)
-
-
-def return_hash_condition(alias: str, _type: str) -> Condition[str]:
-    return Condition(key_alias=alias, key_type=_type, transform=lambda x: abs(hash_unsafe_bytes(x)) % 10)
 
 
 class Conditions:
@@ -68,51 +57,3 @@ class Conditions:
             result_df = self.selector_hash_condition.acts(result_df, single_abi)
 
         return result_df
-
-    @staticmethod
-    def new_datawave_trace_condition():
-        default_trace_function_selector_condition = Condition(
-            key_alias='selector',
-            key_type='string',
-            # :paras x: function abi dict
-            transform=lambda x: encode_hex(function_abi_to_4byte_selector(x))
-        )
-
-        default_trace_function_hash_condition = Condition(
-            key_alias='selector_hash',
-            key_type='int',
-            # :paras x: function abi dict
-            transform=lambda x: abs(hash_unsafe_bytes(encode_hex(function_abi_to_4byte_selector(x)))) % 10
-        )
-
-        return Conditions(
-            ctype='function',
-            address_condition=return_self_condition('to_address', 'string'),
-            address_hash_condition=return_hash_condition('address_hash', 'int'),
-            selector_condition=default_trace_function_selector_condition,
-            selector_hash_condition=default_trace_function_hash_condition
-        )
-
-    @staticmethod
-    def new_datawave_log_condition():
-        default_log_event_selector_condition = Condition(
-            key_alias='selector',
-            key_type='string',
-            # :paras x: event abi dict
-            transform=lambda x: encode_hex(event_abi_to_log_topic(x))
-        )
-
-        default_log_event_hash_condition = Condition(
-            key_alias='selector_hash',
-            key_type='int',
-            # :paras x: event abi dict
-            transform=lambda x: abs(hash_unsafe_bytes(encode_hex(event_abi_to_log_topic(x)))) % 10
-        )
-
-        return Conditions(
-            ctype='event',
-            address_condition=return_self_condition('address', 'string'),
-            address_hash_condition=return_hash_condition('address_hash', 'int'),
-            selector_condition=default_log_event_selector_condition,
-            selector_hash_condition=default_log_event_hash_condition
-        )
